+++ cypress/integration/2.7-marketplace-to-notify-and-validate.spec.js
@@ -1,6 +1,5 @@
 import { loginViaCookies, preserveCookie } from '../support/common/support'
 import { testCase1 } from '../support/outputvalidation'
-import { triggerBroadCaster } from '../support/broadcaster.api'
 import {
   subscribeToProductAlerts,
   verifyEmail,
@@ -8,20 +7,16 @@ import {
 import {
   updateProductStatus,
   updateAppSettings,
-  configureBroadcasterAdapter,
   generateEmailId,
 } from '../support/availability-notify.apis'
 
 const { name, product, warehouseId, skuId } = testCase1
-const workspace = Cypress.env().workspace.name
 const prefix = 'Marketplace to notify'
 
 describe('Testing market place to notify', () => {
   loginViaCookies()
 
   const email = generateEmailId()
-
-  configureBroadcasterAdapter(prefix, workspace)
 
   updateAppSettings(prefix, true)
 
@@ -32,8 +27,6 @@ describe('Testing market place to notify', () => {
   updateAppSettings(prefix, false)
 
   updateProductStatus({ prefix, warehouseId, skuId, unlimited: true })
-
-  triggerBroadCaster(prefix, skuId)
 
   verifyEmail(prefix)