+++ cypress/integration/2.8-availability-notify-validate.spec.js
@@ -1,6 +1,5 @@
 import { loginViaCookies, preserveCookie } from '../support/common/support'
 import { testCase2 } from '../support/outputvalidation'
-import { triggerBroadCaster } from '../support/broadcaster.api'
 import {
   subscribeToProductAlerts,
   verifyEmail,
@@ -8,13 +7,11 @@ import {
 import {
   updateProductStatus,
   updateAppSettings,
-  configureBroadcasterAdapter,
   generateEmailId,
 } from '../support/availability-notify.apis'
 import availbalityNotifyProducts from '../support/products'
 
 const { name, warehouseId, skuId } = testCase2
-const workspace = Cypress.env().workspace.name
 const prefix = 'Availability notify'
 
 describe('Test availability notify scenarios', () => {
@@ -33,13 +30,9 @@ describe('Test availability notify scenarios', () => {
     name,
   })
 
-  configureBroadcasterAdapter(prefix, workspace)
-
   updateAppSettings(prefix, false)
 
   updateProductStatus({ prefix, warehouseId, skuId, unlimited: true })
-
-  triggerBroadCaster(prefix, skuId)
 
   verifyEmail(prefix)